backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
psycopg2-binary
alembic
python-dotenv 